Update trait values version one
This is some corrections to trait values based on a list of all trait values I have updated in 2018. This list can be added to and the code rerun to continue updating typos and other inconsistences.
This commit is relevant to issue #101 I believe all upper-case letters have been corrected at this stage.