Merging traits
I noticed "seed_diameter" was only used in one study and this variable would be the same as "seed_width". Similarly, "seed_depth" was only used in one study and is the same as the "seed_breadth" trait. I talked with Rachael about this and she agreed we could merge things.

It could also be worth merging "dispersal_appendage" (2 studies) with "dispersal_syndrome" (18 studies).