Correct errors in flowering/fruiting time for #109
Some values are coming back as "NA", because in the
conversion of flowering time data from character to binary, the substitution (from the metadata.yml) file is being applied after the custom R code. The solution there here is to edit the raw data files so that the custom R code works as intended.