Kew_2010, Kew_2012: removing studies
Kew_2012 dataset includes a small number of genus-level seed weight measurements and is being removed.
Kew_2010 is being replaced with Kew_2019_1. The small number of values in Kew_2010 that are not in Kew_2019_1 will be merged in. A copy of the original Kew_2010 data.csv file is in the Kew_2019_1 raw data folder.