Adding missing families to species list.
Added missing families to nearly 300 entries in the species list that were missing them.

There were two mispelled genera on the list. One I've fixed in Schmidt_2003 and the other is 'Symphonemia montana' which should be 'Symphionema montanum.I can't find any studies that use the mis-spelling and assume it is on a branch.