Change "unplaced" to "unplaced for APC"
for load_taxonomic_resources(), the second if statement doesn't work if there's an object called "taxonomic resources" already in the global environment. I've added in a line to remove taxonomic resources but it then makes the if statement a bit redundant as it will load taxonomic resources every time now.