Allow additional columns to be read in to the traits table
modify `parse_data` and `load_study` to allow additional columns such as `collection_type` and `sample_age_class` to be read into the traits table. This allows the data to be read in at either the traits or dataset level.

Will be adding additional code to allow data to be read in at the site level