Reviewed fire studies Sophie's entered
Reviewed Williams_2012, Clarke_2009 and Campbell_2006 - they all look really good. I made a few small metadata changes, to align site properties with those used in other datasets.

Biggest change was mapping in more traits from Clarke_2009, after determining that it had a number of species not duplicated in Clark_2015, which we hadn't expected to be the case.