rework context: 1st examples of new setup for context
These studies are the first (of ~60) with reworked context descriptions, where contextual properties are aligned across datasets, there is no longer a contextual property `description`, and type (field, experiment) is now in square brackets at the end of the description.

These are examples and if they look good I'll realign the rest of the studies.