Harmonizing collection_date for `unknown` studies
For studies where collection_date is `unknown`, even to the resolution of a year, I've now applied the format `unknown/publication year`.

Previously these were a broad mix of formats.

For botanical studies (i.e. floras, taxonomic descriptions) this is absolutely appropriate, because data will have been incorporated into the description from some long-ago unknown time to present. For field studies, data are obviously collected at a certain point in time, but this way there is at least some sort of a date-stamp on the dataset.