Fixed problem with parsing ID
Previously, if there was no column with `individual_id` the function didn't have a "column of unique values" as is required for the new `process_generate_id` function. I've now generated such a column - simply row numbers.

The number of unique `individual_id` & `observation_id` values is now identical for the old & new datasets. Some columns are sorting differently, so the numbering doesn't match perfectly (for ~1/4 of studies), but lots of spot checks indicate that all the values are correctly grouped.