WAH studies
* add newly scraped data related to plant growth form traits
* align authors across all WAH studies

There are 112 species with `plant_growth_form` in WAH_1998 that are not also in one of the WAH_2022 versions. This is a mix of current and non-current taxon names. Wondering how best to deal with these - possibly add a filter in custom R code to only retain these names? (There are 2600+ names only in the new versions, so a giant expansion of coverage.)

```
austraits$traits %>%
  filter(dataset_id %in% c("WAH_1998", "WAH_2022_1", "WAH_2022_2")) %>%
  filter(trait_name == "plant_growth_form") %>%
  select(taxon_name, dataset_id, value) %>%
  group_by(taxon_name, dataset_id) %>%
  mutate(value = paste(value, collapse = " ")) %>%
  ungroup() %>%
  distinct() %>%
  group_by(taxon_name) %>%
  pivot_wider(names_from = dataset_id, values_from = value) %>%
  filter(is.na(WAH_2022_1)&is.na(WAH_2022_2))
```