fix duplication in taxa table
In instances where the `accepted_name` had a different taxon rank to a `synonym` (or other status), the taxon was being duplicated in austraits$taxa with incorrect information filled into binomial and trinomial for one of the variants. Fix is to match taxon_rank to the taxon_rank of the `accepted_name`.