Add contexts to floras so they pivot wider
In the floras there are often replicate trait values for certain taxa, because the same "trait" is measured on slightly different entities (primary leaflets vs secondary leaflets). This information had been read in as measurement remarks, but is now a context. The values are poorly standardised - but not something we can do right now.