Portelli_2023, add new study (#803)
This study has been on a branch as "Vesk_2023" for 2 years, waiting for the manuscript to be published. 

Because of taxonomy list mess in September 2023, took data/metadata files for the study and moved them to a new branch. Re-ran tests, looked at report, confirming everything working.

Report reviewed by Peter Vesk long ago.