Hydraulic datasets (#827)
Merging in hydraulics datasets that have been reviewed by contributor (Blackman_2019, Blackman_2023) or are more than 5 years old AND I've never received a response from the data contributor. For a few of these it is possible there are individual-level datasets that we might get in the future, in which case we'll replace the data file.

Also some updating of units to include "moles of what".