Create taxon_list_allow_replacing.csv
On this branch, another version of the taxon_list, allowing replacements - and hence documenting taxa that had been accepted and are now synonyms. The new name isn't even added unless there is a new dataset using that name, because the old name is still indicated as "accepted" so the algorithm doesn't hunt for an "accepted" name to match to.

This file to be remove from this branch after discussing