Add specimen identifiers (#832)
Retrofit 3 AusTraits studies to include specimen identifiers. These are all datasets from Currency Creek Arboretum where the authors have indicated the Dean Nicolle tree number on which measurements were made.

This is also a confirmation that the newly released version of austraits and the soon-to-be-released version of traits.build are interacting properly with austraits.build, allowing specimen identifers to be added and then output as an additional data table.